added auth middleware to goal api endpoints

+++ controllers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 // @desc Get goals
 // @route GET /api/goals
 // @access private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
     res.status(200).json(goals)
 })
 
@@ -20,10 +21,11 @@ const setGoal = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
-    res.status(201).json({ message: "goal created", goal })
+    res.status(201).json({ goal })
 })
 
 // @desc Update goal
@@ -35,6 +37,20 @@ const updateGoal = asyncHandler(async (req, res) => {
     if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    //check for user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found!')
+    }
+
+    //make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized!')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(goal, req.body, {
@@ -53,6 +69,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    //check for user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found!')
+    }
+
+    //make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized!')
     }
 
     const deletedGoal = await Goal.findByIdAndDelete(goal)
